Change graceful to opt-in to signal handling
Previously, a set of standard signals would be handled automatically via
an init() function, however that made the package difficult to use in
packages in which an HTTP server would only be spawned some of the times
(perhaps keyed on an environment variable or flag). Now, signals must be
registered manually.

By default, the top-level "goji" package automatically registers
signals with graceful, so this will result in no behavior changes for
most people.

Fixes #35.